+++ packages/emoji/src/utils/createEmojisFromStrategy.ts
@@ -1,10 +1,20 @@
-import { emojiList, toShort } from 'emoji-toolkit';
+import { emojiList, toShort, shortnameLookup } from 'emoji-toolkit';
 import data from 'emojibase-data/en/compact.json';
+import { EmojiShape } from 'packages/emoji/src/constants/type';
 
 export interface EmojiStrategy {
   [x: string]: {
-    [x: string]: string[];
+    [x: string]: EmojiShape[];
   };
+}
+
+// `toShort` from the toolkit uses a long regular expression that can perform quite poorly.
+// We leverage their (incorrectly typed) `shortnameLookup` object to short-circuit the lookup for common Emoji.
+function fasterUnicodeToShortname(unicode: string): EmojiShape {
+  const shortname =
+    (shortnameLookup as unknown as { [unicode: string]: string })[unicode] ||
+    toShort(unicode);
+  return { shortname, unicode };
 }
 
 // Filtering out all non printable characters.
@@ -17,19 +27,19 @@ export default function createEmojisFromStrategy(): EmojiStrategy {
   const emojis: EmojiStrategy = {};
 
   for (const item of data) {
-    const shortName = toShort(item.unicode);
-    const emoji = emojiList[escapeNonASCIICharacters(shortName)];
+    const emojiShape = fasterUnicodeToShortname(item.unicode);
+    const emoji = emojiList[escapeNonASCIICharacters(emojiShape.shortname)];
     if (emoji) {
       if (!emojis[emoji.category]) {
         emojis[emoji.category] = {};
       }
-      emojis[emoji.category][shortName] = [shortName];
+      emojis[emoji.category][emojiShape.shortname] = [emojiShape];
 
       if (item.skins) {
         for (const skin of item.skins) {
-          const skinShortName = toShort(skin.unicode);
-          if (emojiList[skinShortName]) {
-            emojis[emoji.category][shortName].push(skinShortName);
+          const skinEmoji = fasterUnicodeToShortname(skin.unicode);
+          if (emojiList[skinEmoji.shortname]) {
+            emojis[emoji.category][emojiShape.shortname].push(skinEmoji);
           }
         }
       }