+++ packages/emoji/CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - change EmojiSuggestions to popper.js with option `suggestionsPopperOptions` and deprecate `positionSuggestions` [#3254](https://github.com/draft-js-plugins/draft-js-plugins/issues/3254)
 - change EmojiSelect to use popper.js with option `selectPopperOptions` [#3260](https://github.com/draft-js-plugins/draft-js-plugins/issues/3260)
+- speed up initial import for native Emoji [#3255](https://github.com/draft-js-plugins/draft-js-plugins/issues/3255)
 
 ## 4.6.6
 - added basic keyboard navigation to fix [#3233](https://github.com/draft-js-plugins/draft-js-plugins/issues/3233)